01_ai_agents_first/03_litellm_openai_agent/02_uv_litellm_chainlite/chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "litellm>=1.73.1",
    "python-dotenv>=1.1.1",
]
